third_party/blink/web_tests/fast/dom/Window/get-set-properties-expected.txt

This page tests getting and setting window properties and functions.


----- tests for getting/setting read-write properties -----

PASS: canGet('Attr') should be 'true' and is.
PASS: canSet('Attr') should be 'true' and is.
PASS: canGet('CDATASection') should be 'true' and is.
PASS: canSet('CDATASection') should be 'true' and is.
PASS: canGet('CSSRule') should be 'true' and is.
PASS: canSet('CSSRule') should be 'true' and is.
PASS: canGet('CSSStyleDeclaration') should be 'true' and is.
PASS: canSet('CSSStyleDeclaration') should be 'true' and is.
PASS: canGet('CharacterData') should be 'true' and is.
PASS: canSet('CharacterData') should be 'true' and is.
PASS: canGet('Comment') should be 'true' and is.
PASS: canSet('Comment') should be 'true' and is.
PASS: canGet('DOMException') should be 'true' and is.
PASS: canSet('DOMException') should be 'true' and is.
PASS: canGet('DOMImplementation') should be 'true' and is.
PASS: canSet('DOMImplementation') should be 'true' and is.
PASS: canGet('DOMParser') should be 'true' and is.
PASS: canSet('DOMParser') should be 'true' and is.
PASS: canGet('Document') should be 'true' and is.
PASS: canSet('Document') should be 'true' and is.
PASS: canGet('DocumentFragment') should be 'true' and is.
PASS: canSet('DocumentFragment') should be 'true' and is.
PASS: canGet('DocumentType') should be 'true' and is.
PASS: canSet('DocumentType') should be 'true' and is.
PASS: canGet('Element') should be 'true' and is.
PASS: canSet('Element') should be 'true' and is.
PASS: canGet('EvalError') should be 'true' and is.
PASS: canSet('EvalError') should be 'true' and is.
PASS: canGet('Event') should be 'true' and is.
PASS: canSet('Event') should be 'true' and is.
PASS: canGet('HTMLAnchorElement') should be 'true' and is.
PASS: canSet('HTMLAnchorElement') should be 'true' and is.
PASS: canGet('HTMLAreaElement') should be 'true' and is.
PASS: canSet('HTMLAreaElement') should be 'true' and is.
PASS: canGet('HTMLBRElement') should be 'true' and is.
PASS: canSet('HTMLBRElement') should be 'true' and is.
PASS: canGet('HTMLBaseElement') should be 'true' and is.
PASS: canSet('HTMLBaseElement') should be 'true' and is.
PASS: canGet('HTMLBodyElement') should be 'true' and is.
PASS: canSet('HTMLBodyElement') should be 'true' and is.
PASS: canGet('HTMLButtonElement') should be 'true' and is.
PASS: canSet('HTMLButtonElement') should be 'true' and is.
PASS: canGet('HTMLCanvasElement') should be 'true' and is.
PASS: canSet('HTMLCanvasElement') should be 'true' and is.
PASS: canGet('HTMLDListElement') should be 'true' and is.
PASS: canSet('HTMLDListElement') should be 'true' and is.
PASS: canGet('HTMLDirectoryElement') should be 'true' and is.
PASS: canSet('HTMLDirectoryElement') should be 'true' and is.
PASS: canGet('HTMLDivElement') should be 'true' and is.
PASS: canSet('HTMLDivElement') should be 'true' and is.
PASS: canGet('HTMLDocument') should be 'true' and is.
PASS: canSet('HTMLDocument') should be 'true' and is.
PASS: canGet('HTMLElement') should be 'true' and is.
PASS: canSet('HTMLElement') should be 'true' and is.
PASS: canGet('HTMLFieldSetElement') should be 'true' and is.
PASS: canSet('HTMLFieldSetElement') should be 'true' and is.
PASS: canGet('HTMLFontElement') should be 'true' and is.
PASS: canSet('HTMLFontElement') should be 'true' and is.
PASS: canGet('HTMLFormElement') should be 'true' and is.
PASS: canSet('HTMLFormElement') should be 'true' and is.
PASS: canGet('HTMLFrameElement') should be 'true' and is.
PASS: canSet('HTMLFrameElement') should be 'true' and is.
PASS: canGet('HTMLFrameSetElement') should be 'true' and is.
PASS: canSet('HTMLFrameSetElement') should be 'true' and is.
PASS: canGet('HTMLHRElement') should be 'true' and is.
PASS: canSet('HTMLHRElement') should be 'true' and is.
PASS: canGet('HTMLHeadElement') should be 'true' and is.
PASS: canSet('HTMLHeadElement') should be 'true' and is.
PASS: canGet('HTMLHeadingElement') should be 'true' and is.
PASS: canSet('HTMLHeadingElement') should be 'true' and is.
PASS: canGet('HTMLHtmlElement') should be 'true' and is.
PASS: canSet('HTMLHtmlElement') should be 'true' and is.
PASS: canGet('HTMLIFrameElement') should be 'true' and is.
PASS: canSet('HTMLIFrameElement') should be 'true' and is.
PASS: canGet('HTMLImageElement') should be 'true' and is.
PASS: canSet('HTMLImageElement') should be 'true' and is.
PASS: canGet('HTMLInputElement') should be 'true' and is.
PASS: canSet('HTMLInputElement') should be 'true' and is.
*** FAIL: canGet('HTMLIsIndexElement') should be 'true' but instead is false. ***
PASS: canSet('HTMLIsIndexElement') should be 'true' and is.
PASS: canGet('HTMLLIElement') should be 'true' and is.
PASS: canSet('HTMLLIElement') should be 'true' and is.
PASS: canGet('HTMLLabelElement') should be 'true' and is.
PASS: canSet('HTMLLabelElement') should be 'true' and is.
PASS: canGet('HTMLLegendElement') should be 'true' and is.
PASS: canSet('HTMLLegendElement') should be 'true' and is.
PASS: canGet('HTMLLinkElement') should be 'true' and is.
PASS: canSet('HTMLLinkElement') should be 'true' and is.
PASS: canGet('HTMLMapElement') should be 'true' and is.
PASS: canSet('HTMLMapElement') should be 'true' and is.
PASS: canGet('HTMLMarqueeElement') should be 'true' and is.
PASS: canSet('HTMLMarqueeElement') should be 'true' and is.
PASS: canGet('HTMLMenuElement') should be 'true' and is.
PASS: canSet('HTMLMenuElement') should be 'true' and is.
PASS: canGet('HTMLMetaElement') should be 'true' and is.
PASS: canSet('HTMLMetaElement') should be 'true' and is.
PASS: canGet('HTMLModElement') should be 'true' and is.
PASS: canSet('HTMLModElement') should be 'true' and is.
PASS: canGet('HTMLOListElement') should be 'true' and is.
PASS: canSet('HTMLOListElement') should be 'true' and is.
PASS: canGet('HTMLOptGroupElement') should be 'true' and is.
PASS: canSet('HTMLOptGroupElement') should be 'true' and is.
PASS: canGet('HTMLOptionElement') should be 'true' and is.
PASS: canSet('HTMLOptionElement') should be 'true' and is.
PASS: canGet('HTMLParagraphElement') should be 'true' and is.
PASS: canSet('HTMLParagraphElement') should be 'true' and is.
PASS: canGet('HTMLParamElement') should be 'true' and is.
PASS: canSet('HTMLParamElement') should be 'true' and is.
PASS: canGet('HTMLPreElement') should be 'true' and is.
PASS: canSet('HTMLPreElement') should be 'true' and is.
PASS: canGet('HTMLQuoteElement') should be 'true' and is.
PASS: canSet('HTMLQuoteElement') should be 'true' and is.
PASS: canGet('HTMLScriptElement') should be 'true' and is.
PASS: canSet('HTMLScriptElement') should be 'true' and is.
PASS: canGet('HTMLSelectElement') should be 'true' and is.
PASS: canSet('HTMLSelectElement') should be 'true' and is.
PASS: canGet('HTMLStyleElement') should be 'true' and is.
PASS: canSet('HTMLStyleElement') should be 'true' and is.
PASS: canGet('HTMLTableCaptionElement') should be 'true' and is.
PASS: canSet('HTMLTableCaptionElement') should be 'true' and is.
PASS: canGet('HTMLTableCellElement') should be 'true' and is.
PASS: canSet('HTMLTableCellElement') should be 'true' and is.
PASS: canGet('HTMLTableColElement') should be 'true' and is.
PASS: canSet('HTMLTableColElement') should be 'true' and is.
PASS: canGet('HTMLTableElement') should be 'true' and is.
PASS: canSet('HTMLTableElement') should be 'true' and is.
PASS: canGet('HTMLTableRowElement') should be 'true' and is.
PASS: canSet('HTMLTableRowElement') should be 'true' and is.
PASS: canGet('HTMLTableSectionElement') should be 'true' and is.
PASS: canSet('HTMLTableSectionElement') should be 'true' and is.
PASS: canGet('HTMLTextAreaElement') should be 'true' and is.
PASS: canSet('HTMLTextAreaElement') should be 'true' and is.
PASS: canGet('HTMLTitleElement') should be 'true' and is.
PASS: canSet('HTMLTitleElement') should be 'true' and is.
PASS: canGet('HTMLUListElement') should be 'true' and is.
PASS: canSet('HTMLUListElement') should be 'true' and is.
PASS: canGet('Image') should be 'true' and is.
PASS: canSet('Image') should be 'true' and is.
PASS: canGet('MutationEvent') should be 'true' and is.
PASS: canSet('MutationEvent') should be 'true' and is.
PASS: canGet('Node') should be 'true' and is.
PASS: canSet('Node') should be 'true' and is.
PASS: canGet('NodeFilter') should be 'true' and is.
PASS: canSet('NodeFilter') should be 'true' and is.
PASS: canGet('Option') should be 'true' and is.
PASS: canSet('Option') should be 'true' and is.
PASS: canGet('ProcessingInstruction') should be 'true' and is.
PASS: canSet('ProcessingInstruction') should be 'true' and is.
PASS: canGet('Range') should be 'true' and is.
PASS: canSet('Range') should be 'true' and is.
PASS: canGet('RangeError') should be 'true' and is.
PASS: canSet('RangeError') should be 'true' and is.
PASS: canGet('ReferenceError') should be 'true' and is.
PASS: canSet('ReferenceError') should be 'true' and is.
PASS: canGet('SyntaxError') should be 'true' and is.
PASS: canSet('SyntaxError') should be 'true' and is.
PASS: canGet('Text') should be 'true' and is.
PASS: canSet('Text') should be 'true' and is.
PASS: canGet('TypeError') should be 'true' and is.
PASS: canSet('TypeError') should be 'true' and is.
PASS: canGet('URIError') should be 'true' and is.
PASS: canSet('URIError') should be 'true' and is.
PASS: canGet('XMLDocument') should be 'true' and is.
PASS: canSet('XMLDocument') should be 'true' and is.
PASS: canGet('XMLHttpRequest') should be 'true' and is.
PASS: canSet('XMLHttpRequest') should be 'true' and is.
PASS: canGet('XMLSerializer') should be 'true' and is.
PASS: canSet('XMLSerializer') should be 'true' and is.
PASS: canGet('XPathEvaluator') should be 'true' and is.
PASS: canSet('XPathEvaluator') should be 'true' and is.
PASS: canGet('XPathResult') should be 'true' and is.
PASS: canSet('XPathResult') should be 'true' and is.
PASS: canGet('XSLTProcessor') should be 'true' and is.
PASS: canSet('XSLTProcessor') should be 'true' and is.
PASS: canGet('addEventListener') should be 'true' and is.
PASS: canSet('addEventListener') should be 'true' and is.
PASS: canGet('alert') should be 'true' and is.
PASS: canSet('alert') should be 'true' and is.
PASS: canGet('atob') should be 'true' and is.
PASS: canSet('atob') should be 'true' and is.
PASS: canGet('btoa') should be 'true' and is.
PASS: canSet('btoa') should be 'true' and is.
PASS: canGet('captureEvents') should be 'true' and is.
PASS: canSet('captureEvents') should be 'true' and is.
PASS: canGet('clearInterval') should be 'true' and is.
PASS: canSet('clearInterval') should be 'true' and is.
PASS: canGet('clearTimeout') should be 'true' and is.
PASS: canSet('clearTimeout') should be 'true' and is.
PASS: canGet('clientInformation') should be 'true' and is.
PASS: canSet('clientInformation') should be 'true' and is.
PASS: canGet('console') should be 'true' and is.
PASS: canSet('console') should be 'true' and is.
PASS: canGet('devicePixelRatio') should be 'true' and is.
PASS: canSet('devicePixelRatio') should be 'true' and is.
PASS: canGet('event') should be 'true' and is.
PASS: canSet('event') should be 'true' and is.
PASS: canGet('frames') should be 'true' and is.
PASS: canSet('frames') should be 'true' and is.
PASS: canGet('innerHeight') should be 'true' and is.
PASS: canSet('innerHeight') should be 'true' and is.
PASS: canGet('innerWidth') should be 'true' and is.
PASS: canSet('innerWidth') should be 'true' and is.
PASS: canGet('length') should be 'true' and is.
PASS: canSet('length') should be 'true' and is.
PASS: canGet('locationbar') should be 'true' and is.
PASS: canSet('locationbar') should be 'true' and is.
PASS: canGet('menubar') should be 'true' and is.
PASS: canSet('menubar') should be 'true' and is.
PASS: canGet('offscreenBuffering') should be 'true' and is.
PASS: canSet('offscreenBuffering') should be 'true' and is.
PASS: canGet('opener') should be 'true' and is.
PASS: canSet('opener') should be 'true' and is.
PASS: canGet('outerHeight') should be 'true' and is.
PASS: canSet('outerHeight') should be 'true' and is.
PASS: canGet('outerWidth') should be 'true' and is.
PASS: canSet('outerWidth') should be 'true' and is.
PASS: canGet('pageXOffset') should be 'true' and is.
PASS: canSet('pageXOffset') should be 'true' and is.
PASS: canGet('pageYOffset') should be 'true' and is.
PASS: canSet('pageYOffset') should be 'true' and is.
PASS: canGet('parent') should be 'true' and is.
PASS: canSet('parent') should be 'true' and is.
PASS: canGet('personalbar') should be 'true' and is.
PASS: canSet('personalbar') should be 'true' and is.
PASS: canGet('screen') should be 'true' and is.
PASS: canSet('screen') should be 'true' and is.
PASS: canGet('screenLeft') should be 'true' and is.
PASS: canSet('screenLeft') should be 'true' and is.
PASS: canGet('screenTop') should be 'true' and is.
PASS: canSet('screenTop') should be 'true' and is.
PASS: canGet('screenX') should be 'true' and is.
PASS: canSet('screenX') should be 'true' and is.
PASS: canGet('screenY') should be 'true' and is.
PASS: canSet('screenY') should be 'true' and is.
PASS: canGet('scrollX') should be 'true' and is.
PASS: canSet('scrollX') should be 'true' and is.
PASS: canGet('scrollY') should be 'true' and is.
PASS: canSet('scrollY') should be 'true' and is.
PASS: canGet('scrollbars') should be 'true' and is.
PASS: canSet('scrollbars') should be 'true' and is.
PASS: canGet('self') should be 'true' and is.
PASS: canSet('self') should be 'true' and is.
PASS: canGet('statusbar') should be 'true' and is.
PASS: canSet('statusbar') should be 'true' and is.
PASS: canGet('toolbar') should be 'true' and is.
PASS: canSet('toolbar') should be 'true' and is.

----- tests for getting/setting readonly properties -----

PASS: canGet('closed') should be 'true' and is.
PASS: canSet('closed') should be 'false' and is.
PASS: canGet('document') should be 'true' and is.
PASS: canSet('document') should be 'false' and is.
PASS: canGet('history') should be 'true' and is.
PASS: canSet('history') should be 'false' and is.
PASS: canGet('navigator') should be 'true' and is.
PASS: canSet('navigator') should be 'false' and is.
PASS: canGet('window') should be 'true' and is.
PASS: canSet('window') should be 'false' and is.
PASS: canGet('top') should be 'true' and is.
PASS: canSet('top') should be 'false' and is.

----- tests for getting/setting string properties -----

PASS: canGet('name') should be 'true' and is.
PASS: canSetWithTypeConversion('name') should be 'true' and is.
PASS: canGet('status') should be 'true' and is.
PASS: canSetWithTypeConversion('status') should be 'true' and is.

----- tests for getting/setting functions -----

PASS: canGet('blur') should be 'true' and is.
PASS: canSet('blur') should be 'true' and is.
PASS: canGet('close') should be 'true' and is.
PASS: canSet('close') should be 'true' and is.
PASS: canGet('confirm') should be 'true' and is.
PASS: canSet('confirm') should be 'true' and is.
PASS: canGet('find') should be 'true' and is.
PASS: canSet('find') should be 'true' and is.
PASS: canGet('focus') should be 'true' and is.
PASS: canSet('focus') should be 'true' and is.
PASS: canGet('getComputedStyle') should be 'true' and is.
PASS: canSet('getComputedStyle') should be 'true' and is.
PASS: canGet('getSelection') should be 'true' and is.
PASS: canSet('getSelection') should be 'true' and is.
PASS: canGet('moveBy') should be 'true' and is.
PASS: canSet('moveBy') should be 'true' and is.
PASS: canGet('moveTo') should be 'true' and is.
PASS: canSet('moveTo') should be 'true' and is.
PASS: canGet('open') should be 'true' and is.
PASS: canSet('open') should be 'true' and is.
PASS: canGet('print') should be 'true' and is.
PASS: canSet('print') should be 'true' and is.
PASS: canGet('prompt') should be 'true' and is.
PASS: canSet('prompt') should be 'true' and is.
PASS: canGet('releaseEvents') should be 'true' and is.
PASS: canSet('releaseEvents') should be 'true' and is.
PASS: canGet('removeEventListener') should be 'true' and is.
PASS: canSet('removeEventListener') should be 'true' and is.
PASS: canGet('resizeBy') should be 'true' and is.
PASS: canSet('resizeBy') should be 'true' and is.
PASS: canGet('resizeTo') should be 'true' and is.
PASS: canSet('resizeTo') should be 'true' and is.
PASS: canGet('scroll') should be 'true' and is.
PASS: canSet('scroll') should be 'true' and is.
PASS: canGet('scrollBy') should be 'true' and is.
PASS: canSet('scrollBy') should be 'true' and is.
PASS: canGet('scrollTo') should be 'true' and is.
PASS: canSet('scrollTo') should be 'true' and is.
PASS: canGet('setInterval') should be 'true' and is.
PASS: canSet('setInterval') should be 'true' and is.
PASS: canGet('setTimeout') should be 'true' and is.
PASS: canSet('setTimeout') should be 'true' and is.
PASS: canGet('stop') should be 'true' and is.
PASS: canSet('stop') should be 'true' and is.

----- tests for getting/setting event handlers -----

PASS: canGet('onabort') should be 'true' and is.
PASS: canSetWithCallable('onabort') should be 'true' and is.
PASS: canGet('onbeforeunload') should be 'true' and is.
PASS: canSetWithCallable('onbeforeunload') should be 'true' and is.
PASS: canGet('onblur') should be 'true' and is.
PASS: canSetWithCallable('onblur') should be 'true' and is.
PASS: canGet('onchange') should be 'true' and is.
PASS: canSetWithCallable('onchange') should be 'true' and is.
PASS: canGet('onclick') should be 'true' and is.
PASS: canSetWithCallable('onclick') should be 'true' and is.
PASS: canGet('ondblclick') should be 'true' and is.
PASS: canSetWithCallable('ondblclick') should be 'true' and is.
PASS: canGet('onerror') should be 'true' and is.
PASS: canSetWithCallable('onerror') should be 'true' and is.
PASS: canGet('onfocus') should be 'true' and is.
PASS: canSetWithCallable('onfocus') should be 'true' and is.
PASS: canGet('onkeydown') should be 'true' and is.
PASS: canSetWithCallable('onkeydown') should be 'true' and is.
PASS: canGet('onkeypress') should be 'true' and is.
PASS: canSetWithCallable('onkeypress') should be 'true' and is.
PASS: canGet('onkeyup') should be 'true' and is.
PASS: canSetWithCallable('onkeyup') should be 'true' and is.
PASS: canGet('onload') should be 'true' and is.
PASS: canSetWithCallable('onload') should be 'true' and is.
PASS: canGet('onmousedown') should be 'true' and is.
PASS: canSetWithCallable('onmousedown') should be 'true' and is.
PASS: canGet('onmousemove') should be 'true' and is.
PASS: canSetWithCallable('onmousemove') should be 'true' and is.
PASS: canGet('onmouseout') should be 'true' and is.
PASS: canSetWithCallable('onmouseout') should be 'true' and is.
PASS: canGet('onmouseover') should be 'true' and is.
PASS: canSetWithCallable('onmouseover') should be 'true' and is.
PASS: canGet('onmouseup') should be 'true' and is.
PASS: canSetWithCallable('onmouseup') should be 'true' and is.
PASS: canGet('onmousewheel') should be 'true' and is.
PASS: canSetWithCallable('onmousewheel') should be 'true' and is.
PASS: canGet('onreset') should be 'true' and is.
PASS: canSetWithCallable('onreset') should be 'true' and is.
PASS: canGet('onresize') should be 'true' and is.
PASS: canSetWithCallable('onresize') should be 'true' and is.
PASS: canGet('onscroll') should be 'true' and is.
PASS: canSetWithCallable('onscroll') should be 'true' and is.
PASS: canGet('onsearch') should be 'true' and is.
PASS: canSetWithCallable('onsearch') should be 'true' and is.
PASS: canGet('onselect') should be 'true' and is.
PASS: canSetWithCallable('onselect') should be 'true' and is.
PASS: canGet('onsubmit') should be 'true' and is.
PASS: canSetWithCallable('onsubmit') should be 'true' and is.
PASS: canGet('onunload') should be 'true' and is.
PASS: canSetWithCallable('onunload') should be 'true' and is.
